# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Production ready starter guide.",
        "title": "My Go API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/cycle-items/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cycle_items"
                ],
                "summary": "Get a cycle item by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.CycleItemResponse"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cycle_items"
                ],
                "summary": "Update a cycle item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Cycle item info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.UpdateCycleItemRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "cycle_items"
                ],
                "summary": "Delete a cycle item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/exercise-logs": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exercise_logs"
                ],
                "summary": "Create a new exercise log",
                "parameters": [
                    {
                        "description": "Exercise log info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateExerciseLogRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.ExerciseLogResponse"
                        }
                    }
                }
            }
        },
        "/exercise-logs/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exercise_logs"
                ],
                "summary": "Get an exercise log by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Log ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.ExerciseLogResponse"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "exercise_logs"
                ],
                "summary": "Delete an exercise log",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Log ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
        "/study-cycles/{id}/items": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cycle_items"
                ],
                "summary": "List all items for a cycle",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cycle ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.CycleItemResponse"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cycle_items"
                ],
                "summary": "Create a new cycle item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cycle ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Cycle item info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateCycleItemRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.CycleItemResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.CreateCycleItemRequest": {
            "type": "object",
            "required": [
                "order_index",
                "subject_id"
            ],
            "properties": {
                "order_index": {
                    "type": "integer",
                    "minimum": 1
                },
                "planned_duration_minutes": {
                    "type": "integer",
                    "minimum": 1
                },
                "subject_id": {
                    "type": "string"
                }
            }
        },
        "handler.CreateExerciseLogRequest": {
            "type": "object",
            "required": [
                "correct_count",
                "questions_count",
                "subject_id"
            ],
            "properties": {
                "correct_count": {
                    "type": "integer",
                    "minimum": 0
                },
                "questions_count": {
                    "type": "integer",
                    "minimum": 0
                },
                "session_id": {
                    "type": "string"
                },
                "subject_id": {
                    "type": "string"
                },
                "topic_id": {
                    "type": "string"
                }
            }
        },
...
        "handler.CycleItemResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "cycle_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "order_index": {
                    "type": "integer"
                },
                "planned_duration_minutes": {
                    "type": "integer"
                },
                "subject_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "handler.ExerciseLogResponse": {
            "type": "object",
            "properties": {
                "correct_count": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "questions_count": {
                    "type": "integer"
                },
                "session_id": {
                    "type": "string"
                },
                "subject_id": {
                    "type": "string"
                },
                "topic_id": {
                    "type": "string"
                }
            }
        },
...
        "handler.UpdateCycleItemRequest": {
            "type": "object",
            "required": [
                "order_index",
                "subject_id"
            ],
            "properties": {
                "order_index": {
                    "type": "integer",
                    "minimum": 1
                },
                "planned_duration_minutes": {
                    "type": "integer",
                    "minimum": 1
                },
                "subject_id": {
                    "type": "string"
                }
            }
        },
...
    }
}